+++ src/Settings/index.ts
@@ -1,4 +1,4 @@
-import { TestState } from './DebugCodeLens'
+import { TestState } from '../DebugCodeLens'
 
 export interface IPluginSettings {
   autoEnable?: boolean
@@ -16,4 +16,12 @@ export interface IPluginSettings {
   runAllTestsFirst?: boolean
   showCoverageOnLoad: boolean
   coverageFormatter: string
+}
+
+export function isDefaultPathToJest(str) {
+  return str === null || str === ''
+}
+
+export function hasUserSetPathToJest(str) {
+  return !isDefaultPathToJest(str)
 }