+++ tests/helpers.test.ts
@@ -1,65 +1,77 @@
 jest.unmock('../src/helpers')
-jest.mock('fs')
+jest.unmock('../src/Settings')
 
-import { pathToJestPackageJSON, isCRATestCommand } from '../src/helpers'
-import { existsSync } from 'fs'
+const mockExistsSync = jest.fn()
+const mockReadFileSync = jest.fn()
+jest.mock('fs', () => ({
+  existsSync: mockExistsSync,
+  readFileSync: mockReadFileSync,
+}))
+
+const mockPlatform = jest.fn()
+jest.mock('os', () => ({ platform: mockPlatform }))
+
+const mockJoin = jest.fn()
+const mockNormalize = jest.fn()
+jest.mock('path', () => ({
+  join: mockJoin,
+  normalize: mockNormalize,
+}))
+
+import { isCreateReactAppTestCommand, pathToJest, pathToJestPackageJSON } from '../src/helpers'
 import * as path from 'path'
-
-const existsMock = existsSync as jest.Mock<boolean>
-const defaultPathToJest = 'node_modules/.bin/jest'
 
 describe('ModuleHelpers', () => {
   describe('pathToJestPackageJSON', () => {
-    it('should return null when not found', () => {
-      existsMock.mockReturnValueOnce(false).mockReturnValueOnce(false)
+    const defaultPathToJest = null
+    const defaultSettings: any = {
+      pathToJest: defaultPathToJest,
+      rootPath: '',
+    }
 
-      const workspace: any = {
-        rootPath: '',
-        pathToJest: defaultPathToJest,
-      }
-      expect(pathToJestPackageJSON(workspace)).toBe(null)
+    beforeEach(() => {
+      mockJoin.mockImplementation(require.requireActual('path').join)
+      mockNormalize.mockImplementation(require.requireActual('path').normalize)
+    })
+
+    it('should return null when not found', () => {
+      mockExistsSync.mockReturnValue(false)
+
+      expect(pathToJestPackageJSON(defaultSettings)).toBe(null)
     })
 
     describe('pathToJest: <default>', () => {
       describe('rootPath: <default>', () => {
         it('should return package.json when Jest is installed as a dependency', () => {
           const expected = path.join('node_modules', 'jest', 'package.json')
-          existsMock.mockImplementation(path => path === expected)
+          mockExistsSync.mockImplementation(path => path === expected)
 
-          const workspace: any = {
-            rootPath: '',
-            pathToJest: defaultPathToJest,
-          }
-          expect(pathToJestPackageJSON(workspace)).toBe(expected)
+          expect(pathToJestPackageJSON(defaultSettings)).toBe(expected)
         })
 
         it('should return package.json when React Scripts are installed as a dependency', () => {
           const expected = path.join('node_modules', 'react-scripts', 'node_modules', 'jest', 'package.json')
-          existsMock.mockImplementation(path => path === expected)
+          mockExistsSync.mockImplementation(path => path === expected)
 
-          const workspace: any = {
-            rootPath: '',
-            pathToJest: defaultPathToJest,
-          }
-          expect(pathToJestPackageJSON(workspace)).toBe(expected)
+          expect(pathToJestPackageJSON(defaultSettings)).toBe(expected)
         })
       })
 
       describe('rootPath: <string>', () => {
         it('should return package.json when Jest is installed as a dependency', () => {
           const expected = path.join('..', '..', 'node_modules', 'jest', 'package.json')
-          existsMock.mockImplementation(path => path === expected)
+          mockExistsSync.mockImplementation(path => path === expected)
 
           const workspace: any = {
-            rootPath: path.join('..', '..'),
             pathToJest: defaultPathToJest,
+            rootPath: path.join('..', '..'),
           }
           expect(pathToJestPackageJSON(workspace)).toBe(expected)
         })
 
         it('should return package.json when Jest-cli is installed as a dependency', () => {
           const expected = path.join('..', '..', 'node_modules', 'jest-cli', 'package.json')
-          existsMock.mockImplementation(path => path === expected)
+          mockExistsSync.mockImplementation(path => path === expected)
 
           const workspace: any = {
             rootPath: path.join('..', '..'),
@@ -78,7 +90,7 @@ describe('ModuleHelpers', () => {
             'jest',
             'package.json'
           )
-          existsMock.mockImplementation(path => path === expected)
+          mockExistsSync.mockImplementation(path => path === expected)
 
           const workspace: any = {
             rootPath: path.join('..', '..'),
@@ -92,7 +104,7 @@ describe('ModuleHelpers', () => {
     describe('pathToJest: npm test --', () => {
       it('will not find package.json', () => {
         const expected = null
-        existsMock.mockImplementation(path => path === expected)
+        mockExistsSync.mockImplementation(path => path === expected)
 
         const workspace: any = {
           rootPath: '',
@@ -105,7 +117,7 @@ describe('ModuleHelpers', () => {
     describe(`pathToJest: ${path.join('"test scripts"', 'test')}`, () => {
       it('will not find package.json', () => {
         const expected = path.join('"test scripts"', 'test')
-        existsMock.mockImplementation(path => path === expected)
+        mockExistsSync.mockImplementation(path => path === expected)
 
         const workspace: any = {
           rootPath: '',
@@ -116,12 +128,89 @@ describe('ModuleHelpers', () => {
     })
   })
 
-  describe('isCRATestCommand', () => {
+  describe('isCreateReactAppTestCommand', () => {
     it('should return true for CRA', () => {
-      expect(isCRATestCommand('react-scripts test --env=jsdom')).toBeTruthy()
+      expect(isCreateReactAppTestCommand('react-scripts test --env=jsdom')).toBe(true)
     })
+
     it('should return false for other scripts', () => {
-      expect(isCRATestCommand('custom-script')).toBeFalsy()
+      expect(isCreateReactAppTestCommand('custom-script')).toBe(false)
+    })
+  })
+
+  describe('pathToJest', () => {
+    const defaultSettings: any = {
+      pathToJest: null,
+      rootPath: '',
+    }
+
+    it('returns "npm test --" when bootstrapped with create-react-app', () => {
+      mockReadFileSync.mockReturnValueOnce(
+        JSON.stringify({
+          scripts: {
+            test: 'react-scripts test',
+          },
+        })
+      )
+
+      expect(pathToJest(defaultSettings)).toBe('npm test --')
+    })
+
+    it('returns the normalized "pathToJest" setting when set by the user', () => {
+      const expected = {}
+      mockNormalize.mockReturnValueOnce(expected)
+
+      const settings: any = {
+        pathToJest: expected,
+        rootPath: '',
+      }
+
+      expect(pathToJest(settings)).toBe(expected)
+      expect(mockNormalize).toBeCalledWith(settings.pathToJest)
+    })
+
+    it('defaults to "node_modules/.bin/jest.cmd" on Windows when Jest is locally installed', () => {
+      const expected = 'node_modules\\.bin\\jest.cmd'
+
+      mockJoin.mockImplementation(require.requireActual('path').win32.join)
+      mockPlatform.mockReturnValue('win32')
+      mockNormalize.mockImplementationOnce(arg => arg)
+      mockExistsSync.mockImplementation(path => path === expected)
+
+      expect(pathToJest(defaultSettings)).toBe(expected)
+    })
+
+    it('defaults to "jest.cmd" on Windows when Jest is not locally installed', () => {
+      const expected = 'jest.cmd'
+
+      mockJoin.mockImplementation(require.requireActual('path').win32.join)
+      mockPlatform.mockReturnValue('win32')
+      mockNormalize.mockImplementationOnce(arg => arg)
+      mockExistsSync.mockImplementation(_ => false)
+
+      expect(pathToJest(defaultSettings)).toBe(expected)
+    })
+
+    it('defaults to "node_modules/.bin/jest" on a non-Windows OS when Jest is locally installed', () => {
+      const expected = 'node_modules/.bin/jest'
+
+      mockJoin.mockImplementation(require.requireActual('path').posix.join)
+      mockPlatform.mockReturnValue('linux')
+      mockNormalize.mockImplementationOnce(arg => arg)
+      mockExistsSync.mockImplementation(path => path === expected)
+
+      expect(pathToJest(defaultSettings)).toBe(expected)
+    })
+
+    it('defaults to "jest" on a non-Windows OS when Jest is locally installed', () => {
+      const expected = 'jest'
+
+      mockJoin.mockImplementation(require.requireActual('path').posix.join)
+      mockPlatform.mockReturnValue('linux')
+      mockNormalize.mockImplementationOnce(arg => arg)
+      mockExistsSync.mockImplementation(_ => false)
+
+      expect(pathToJest(defaultSettings)).toBe(expected)
     })
   })
 })